Move src/commcare_cloud/{secrets => environment/secrets}l
