+++ src/commcare_cloud/environment/secrets/backends/ansible_vault/tests/expected-generated-variables.yml
